refactor: 시간 기준 적용
- 기준 시간을 정확히 표기합니다.
